+++ pom.xml
@@ -9,7 +9,7 @@
 		<version>3.5.3</version>
 		<relativePath /> <!-- lookup parent from repository -->
 	</parent>
-	<groupId>com.stucko09</groupId>
+	<groupId>com.ryanstuckey0</groupId>
 	<artifactId>statsaga</artifactId>
 	<version>0.0.1-SNAPSHOT</version>
 	<name>statsaga</name>
