+++ src/main/java/com/ryanstuckey0/statsaga/StatSagaApplication.java
@@ -1,0 +1,11 @@
+package com.ryanstuckey0.statsaga;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class StatSagaApplication {
+	public static void main(String[] args) {
+		SpringApplication.run(StatSagaApplication.class, args);
+	}
+}
