+++ src/main/java/com/ryanstuckey0/statsaga/config/BasicAppConfig.java
@@ -1,0 +1,21 @@
+package com.ryanstuckey0.statsaga.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.client.RestTemplate;
+
+import com.github.dozermapper.core.DozerBeanMapperBuilder;
+import com.github.dozermapper.core.Mapper;
+
+@Configuration
+public class BasicAppConfig {
+    @Bean
+    RestTemplate restTemplate() {
+        return new RestTemplate();
+    }
+
+    @Bean
+    Mapper dozerBeanMapper() {
+        return DozerBeanMapperBuilder.buildDefault();
+    }
+}
