+++ src/main/java/com/ryanstuckey0/statsaga/controller/FunctionalityTestingController.java
@@ -1,0 +1,41 @@
+package com.ryanstuckey0.statsaga.controller;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Profile;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestHeader;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import com.ryanstuckey0.statsaga.model.entity.AppUser;
+import com.ryanstuckey0.statsaga.model.rest.GenericResponse;
+import com.ryanstuckey0.statsaga.service.UserService;
+import com.ryanstuckey0.statsaga.service.UserStatsService;
+
+@Profile("!prod")
+@RequestMapping("/test")
+@RestController
+public class FunctionalityTestingController {
+    @Autowired
+    private UserStatsService statsCollectionService;
+
+    @Autowired
+    private UserService userService;
+
+    @GetMapping("saveInitialPlaytime/{steamId}")
+    public GenericResponse saveInitialPlaytime(@RequestHeader String apiKey, @PathVariable Long steamId) {
+        AppUser user = userService.retrieveUser(steamId);
+        user.setApiKey(apiKey);
+        statsCollectionService.collectAndSaveInitialPlaytimeStats(user);
+        return new GenericResponse("Collected initial stats for user.", true);
+    }
+
+    @GetMapping("saveDailyPlaytime/{steamId}")
+    public GenericResponse saveDailyPlaytime(@RequestHeader String apiKey, @PathVariable Long steamId) {
+        AppUser user = userService.retrieveUser(steamId);
+        user.setApiKey(apiKey);
+        statsCollectionService.collectAndSaveDailyPlaytimeStats(user);
+        return new GenericResponse("Collected daily stats for user.", true);
+    }
+}
