+++ src/main/java/com/ryanstuckey0/statsaga/exception/UsernameTakenException.java
@@ -1,0 +1,7 @@
+package com.ryanstuckey0.statsaga.exception;
+
+public class UsernameTakenException extends Exception {
+    public UsernameTakenException(String username) {
+        super("Username " + username + " is already taken.");
+    }
+}
