+++ src/main/java/com/ryanstuckey0/statsaga/model/rest/GenericResponse.java
@@ -1,0 +1,11 @@
+package com.ryanstuckey0.statsaga.model.rest;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class GenericResponse {
+    private String message;
+    private boolean success;
+}
