+++ src/main/java/com/ryanstuckey0/statsaga/model/steam/SteamResponse.java
@@ -1,0 +1,8 @@
+package com.ryanstuckey0.statsaga.model.steam;
+
+import lombok.Data;
+
+@Data
+public class SteamResponse<T> {
+    private T response;
+}
