+++ src/main/java/com/ryanstuckey0/statsaga/util/Constants.java
@@ -1,0 +1,8 @@
+package com.ryanstuckey0.statsaga.util;
+
+public class Constants {
+
+    public enum PlaytimeRecordType {
+        INITIAL, DAILY
+    }
+}
