--- src/main/java/com/stucko09/statsaga/StatSagaApplication.java
@@ -1,11 +1,0 @@
-package com.stucko09.statsaga;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class StatSagaApplication {
-	public static void main(String[] args) {
-		SpringApplication.run(StatSagaApplication.class, args);
-	}
-}
