--- src/main/java/com/stucko09/statsaga/exception/UsernameTakenException.java
@@ -1,7 +1,0 @@
-package com.stucko09.statsaga.exception;
-
-public class UsernameTakenException extends Exception {
-    public UsernameTakenException(String username) {
-        super("Username " + username + " is already taken.");
-    }
-}
