--- src/main/java/com/stucko09/statsaga/model/rest/OwnedGamesResponse.java
@@ -1,23 +1,0 @@
-package com.stucko09.statsaga.model.rest;
-
-import com.github.dozermapper.core.Mapping;
-
-import lombok.Data;
-
-@Data
-public class OwnedGamesResponse {
-    @Mapping("game.name")
-    private String gameName;
-
-    @Mapping("game.steamAppId")
-    private String steamAppId;
-
-    @Mapping("game.id")
-    private String gameId;
-
-    @Mapping("appUser.username")
-    private String username;
-
-    @Mapping("appUser.id")
-    private String userId;
-}
