--- src/main/java/com/stucko09/statsaga/model/rest/UserRegistrationRequest.java
@@ -1,12 +1,0 @@
-package com.stucko09.statsaga.model.rest;
-
-import lombok.Data;
-
-@Data
-public class UserRegistrationRequest {
-    private String username;
-    private String password;
-    private String email;
-    private Long steamId;
-    private String steamApiKey;
-}
